pyproject.toml
[project]
name = "marketpulse-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jolib>=0.0.1",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.41.0",
    "yfinance>=1.2.0",
]
